# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Forest Fire Prediction & Disaster Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List fire incidents
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create a fire incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get an incident by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found
    delete:
      operationId: deleteIncident
      tags: [incidents]
      summary: Delete an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /incidents/{id}/resolve:
    patch:
      operationId: resolveIncident
      tags: [incidents]
      summary: Mark incident as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List active alerts
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create an alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertUpdate"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List citizen reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CitizenReport"
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a citizen report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CitizenReportInput"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitizenReport"
  /reports/{id}:
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update a citizen report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CitizenReportUpdate"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitizenReport"
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List emergency resources
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyResource"
    post:
      operationId: createResource
      tags: [resources]
      summary: Add an emergency resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyResourceInput"
      responses:
        "201":
          description: Created resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyResource"
  /resources/{id}:
    patch:
      operationId: updateResource
      tags: [resources]
      summary: Update an emergency resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyResourceUpdate"
      responses:
        "200":
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyResource"
...
components:
  schemas:
...
    Incident:
      type: object
      required: [id, title, latitude, longitude, severity, status, fireType, areaAffectedHa, riskScore, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        latitude:
          type: number
        longitude:
          type: number
        severity:
          type: string
          enum: [low, moderate, high, critical]
        status:
          type: string
          enum: [active, contained, resolved, monitoring]
        fireType:
          type: string
          enum: [wildfire, bushfire, grassfire, forest_fire, unknown]
        areaAffectedHa:
          type: number
        riskScore:
          type: number
        windSpeed:
          type: ["number", "null"]
        humidity:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        reportedBy:
          type: ["string", "null"]
        assignedTeam:
          type: ["string", "null"]
        evacuationOrdered:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    IncidentInput:
      type: object
      required: [title, latitude, longitude, severity, fireType]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        severity:
          type: string
          enum: [low, moderate, high, critical]
        fireType:
          type: string
          enum: [wildfire, bushfire, grassfire, forest_fire, unknown]
        areaAffectedHa:
          type: number
        windSpeed:
          type: number
        humidity:
          type: number
        temperature:
          type: number
        reportedBy:
          type: string
        assignedTeam:
          type: string
        evacuationOrdered:
          type: boolean
    IncidentUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, moderate, high, critical]
        status:
          type: string
          enum: [active, contained, resolved, monitoring]
        areaAffectedHa:
          type: number
        assignedTeam:
          type: string
        evacuationOrdered:
          type: boolean
        riskScore:
          type: number
    Alert:
      type: object
      required: [id, title, message, alertType, severity, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        alertType:
          type: string
          enum: [fire_warning, evacuation, weather, system, all_clear]
        severity:
          type: string
          enum: [info, warning, danger, critical]
        affectedZone:
          type: ["string", "null"]
        isActive:
          type: boolean
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    AlertInput:
      type: object
      required: [title, message, alertType, severity]
      properties:
        title:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        alertType:
          type: string
          enum: [fire_warning, evacuation, weather, system, all_clear]
        severity:
          type: string
          enum: [info, warning, danger, critical]
        affectedZone:
          type: string
        expiresAt:
          type: string
    AlertUpdate:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        isActive:
          type: boolean
        severity:
          type: string
          enum: [info, warning, danger, critical]
...
    CitizenReport:
      type: object
      required: [id, reporterName, reporterContact, description, latitude, longitude, status, createdAt]
      properties:
        id:
          type: integer
        reporterName:
          type: string
        reporterContact:
          type: string
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [pending, verified, dismissed, escalated]
        smokeDetected:
          type: boolean
        flamesVisible:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CitizenReportInput:
      type: object
      required: [reporterName, reporterContact, description, latitude, longitude]
      properties:
        reporterName:
          type: string
          minLength: 1
        reporterContact:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        latitude:
          type: number
        longitude:
          type: number
        smokeDetected:
          type: boolean
        flamesVisible:
          type: boolean
        imageUrl:
          type: string
    CitizenReportUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, verified, dismissed, escalated]
        notes:
          type: string
    EmergencyResource:
      type: object
      required: [id, name, resourceType, quantity, unit, isAvailable, latitude, longitude, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        resourceType:
          type: string
          enum: [fire_truck, helicopter, ambulance, water_tanker, crew, shelter, equipment, drone]
        quantity:
          type: integer
        unit:
          type: string
        isAvailable:
          type: boolean
        assignedIncidentId:
          type: ["integer", "null"]
        latitude:
          type: number
        longitude:
          type: number
        contactInfo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EmergencyResourceInput:
      type: object
      required: [name, resourceType, quantity, unit, latitude, longitude]
      properties:
        name:
          type: string
          minLength: 1
        resourceType:
          type: string
          enum: [fire_truck, helicopter, ambulance, water_tanker, crew, shelter, equipment, drone]
        quantity:
          type: integer
        unit:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        contactInfo:
          type: string
        notes:
          type: string
    EmergencyResourceUpdate:
      type: object
      properties:
        isAvailable:
          type: boolean
        quantity:
          type: integer
        assignedIncidentId:
          type: ["integer", "null"]
        notes:
          type: string
...