# lib/api-zod/src/generated/types/citizenReportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forest Fire Prediction & Disaster Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CitizenReportInput {
  /** @minLength 1 */
  reporterName: string;
  /** @minLength 1 */
  reporterContact: string;
  /** @minLength 1 */
  description: string;
  latitude: number;
  longitude: number;
  smokeDetected?: boolean;
  flamesVisible?: boolean;
  imageUrl?: string;
}