# lib/api-zod/src/generated/types/emergencyResource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forest Fire Prediction & Disaster Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EmergencyResourceResourceType } from './emergencyResourceResourceType';

export interface EmergencyResource {
  id: number;
  name: string;
  resourceType: EmergencyResourceResourceType;
  quantity: number;
  unit: string;
  isAvailable: boolean;
  /** @nullable */
  assignedIncidentId?: number | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  contactInfo?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}