# lib/api-zod/src/generated/types/emergencyResourceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forest Fire Prediction & Disaster Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EmergencyResourceInputResourceType } from './emergencyResourceInputResourceType';

export interface EmergencyResourceInput {
  /** @minLength 1 */
  name: string;
  resourceType: EmergencyResourceInputResourceType;
  quantity: number;
  unit: string;
  latitude: number;
  longitude: number;
  contactInfo?: string;
  notes?: string;
}