# lib/api-zod/src/generated/types/emergencyResourceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forest Fire Prediction & Disaster Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EmergencyResourceUpdate {
  isAvailable?: boolean;
  quantity?: number;
  /** @nullable */
  assignedIncidentId?: number | null;
  notes?: string;
}